+++ main.go
@@ -31,6 +31,10 @@ func initialize() {
 }
 
 func main() {
+	if len(os.Args) > 1 && os.Args[1] == "health" {
+		healthcheck()
+	}
+
 	initialize()
 	defer shutdownVips()